+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## Under development
+# Under development
+* Improve error messages (#7)
+* Add become/become_user/ignore to play DSL
+* Prevent these Ruby methods from interfering with module attributes: `[:system, :test, :warn, :sleep]` 
+
+# 1.0.3 - 27 August 2016
 * Fix register count issue (#30)
 * Inclusion support from playbooks
 * Handler support